iommu/context: embed the iova map in the iommu context
Embed the iova_map directly into iommu/context and iommu/dma.
