+++ src/iommu/dma.c
@@ -12,17 +12,145 @@
 
 #define log_fmt(fmt) "iommu/dma: " fmt
 
+#include <errno.h>
+#include <inttypes.h>
 #include <stddef.h>
 #include <stdint.h>
 #include <pthread.h>
 
-#include "ccan/list/list.h"
+#include "ccan/minmax/minmax.h"
 
-#include "vfn/iommu.h"
 #include "vfn/support.h"
+#include "vfn/iommu.h"
 
-#include "util/iova_map.h"
 #include "context.h"
+
+static int iova_cmp(const void *vaddr, const struct skiplist_node *n)
+{
+	struct iova_mapping *m = container_of_var(n, m, list);
+
+	if (vaddr < m->vaddr)
+		return -1;
+	else if (vaddr >= m->vaddr + m->len)
+		return 1;
+
+	return 0;
+}
+
+static int iova_map_add(struct iova_map *map, void *vaddr, size_t len, uint64_t iova)
+{
+	__autolock(&map->lock);
+
+	struct skiplist_node *update[SKIPLIST_LEVELS] = {};
+	struct iova_mapping *m;
+
+	if (!len) {
+		errno = EINVAL;
+		return -1;
+	}
+
+	if (skiplist_find(&map->list, vaddr, iova_cmp, update)) {
+		errno = EEXIST;
+		return -1;
+	}
+
+	m = znew_t(struct iova_mapping, 1);
+
+	m->vaddr = vaddr;
+	m->len = len;
+	m->iova = iova;
+
+	skiplist_link(&map->list, &m->list, update);
+
+	return 0;
+}
+
+static void iova_map_remove(struct iova_map *map, void *vaddr)
+{
+	__autolock(&map->lock);
+
+	struct skiplist_node *n, *update[SKIPLIST_LEVELS] = {};
+
+	n = skiplist_find(&map->list, vaddr, iova_cmp, update);
+	if (!n)
+		return;
+
+	skiplist_erase(&map->list, n, update);
+}
+
+static struct iova_mapping *iova_map_find(struct iova_map *map, void *vaddr)
+{
+	__autolock(&map->lock);
+
+	return container_of_or_null(skiplist_find(&map->list, vaddr, iova_cmp, NULL),
+				    struct iova_mapping, list);
+}
+
+static void iova_map_clear_with(struct iova_map *map, skiplist_iter_fn fn, void *opaque)
+{
+	__autolock(&map->lock);
+
+	skiplist_clear_with(&map->list, fn, opaque);
+}
+
+static void iova_map_clear(struct iova_map *map)
+{
+	iova_map_clear_with(map, NULL, NULL);
+}
+
+void iova_map_destroy(struct iova_map *map)
+{
+       iova_map_clear(map);
+
+       free(map->iova_ranges);
+
+       memset(map, 0x0, sizeof(*map));
+}
+
+static bool iova_map_translate(struct iova_map *map, void *vaddr, uint64_t *iova)
+{
+	struct iova_mapping *m = iova_map_find(map, vaddr);
+
+	if (m) {
+		*iova = m->iova + (vaddr - m->vaddr);
+		return true;
+	}
+
+	return false;
+}
+
+int iova_map_reserve(struct iova_map *map, size_t len, uint64_t *iova)
+{
+	__autolock(&map->lock);
+
+	if (!ALIGNED(len, __VFN_PAGESIZE)) {
+		log_debug("len is not page aligned\n");
+		errno = EINVAL;
+		return -1;
+	}
+
+	for (int i = 0; i < map->nranges; i++) {
+		struct iommu_iova_range *r = &map->iova_ranges[i];
+		uint64_t next = map->next;
+
+		if (r->last < next)
+			continue;
+
+		next = max_t(uint64_t, next, r->start);
+
+		if (next > r->last || r->last - next + 1 < len)
+			continue;
+
+		map->next = next + len;
+
+		*iova = next;
+
+		return 0;
+	}
+
+	errno = ENOMEM;
+	return -1;
+}
 
 int iommu_map_vaddr(struct iommu_ctx *ctx, void *vaddr, size_t len, uint64_t *iova,
 		    unsigned long flags)
